+++ pkg/scanners/clair/convert.go
@@ -1,9 +1,9 @@
 package clair
 
 import (
-	clairV1 "github.com/coreos/clair/api/v1"
 	"github.com/stackrox/rox/generated/storage"
 	clairConv "github.com/stackrox/rox/pkg/clair"
+	clairV1 "github.com/stackrox/scanner/api/v1"
 )
 
 func convertLayerToImageScan(image *storage.Image, layerEnvelope *clairV1.LayerEnvelope) *storage.ImageScan {