+++ ui/src/Containers/Images/ImageDetails.js
@@ -154,7 +154,7 @@ const ImageDetails = ({
         <Panel
             header={header}
             onClose={unselectImage}
-            className="bg-primary-200 w-full h-full absolute pin-r pin-t md:w-2/3 min-w-72 md:relative"
+            className="bg-primary-200 w-full h-full absolute pin-r pin-t md:w-2/3 min-w-1/2 md:relative"
         >
             {content}
         </Panel>